Refactor MapResources to own file, package mapgenerator
